wa-qr-ia/ia-backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23
pydantic>=2.0
Pillow>=10.0
pytesseract>=0.3
boto3>=1.28

